@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import 'package:Shrine/model/dispatcher.dart';
+import 'package:shrine/model/dispatcher.dart';
 import 'package:flutter/material.dart';
 
 import '../model/product.dart';
