@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:Shrine/app.dart';
-import 'package:Shrine/model/app_state_model.dart';
+import 'package:shrine/app.dart';
+import 'package:shrine/model/app_state_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:scoped_model/scoped_model.dart';
